+++ daemon/task_handler/messages/pause.rs
@@ -1,6 +1,10 @@
-use super::*;
+use log::{error, info};
+
+use pueue_lib::state::GroupStatus;
+use pueue_lib::task::TaskStatus;
 
 use crate::ok_or_shutdown;
+use crate::task_handler::{LockedState, ProcessAction, TaskHandler};
 
 impl TaskHandler {
     /// Pause specific tasks or groups.